⚡ Add `SVGIcon` component stories
Why
- To add initial visual tests to the `SVGIcon` component

How
- Removed all the `Storybook` template code
- Refactored `SVGIcon` to show a default "missing" icon if the defined icon name doesn't exist
- Added `SVGIcon` stories

Tags
 Changes to be committed:
	new file:   src/components/SVGIcon/SVGIcon.stories.ts
	modified:   src/components/SVGIcon/SVGIcon.tsx
	modified:   src/hooks/useDynamicSVG.ts
	deleted:    src/stories/Button.stories.ts
	deleted:    src/stories/Button.tsx
	deleted:    src/stories/Configure.mdx
	deleted:    src/stories/Header.stories.ts
	deleted:    src/stories/Header.tsx
	deleted:    src/stories/Page.stories.ts
	deleted:    src/stories/Page.tsx
	deleted:    src/stories/assets/accessibility.png
	deleted:    src/stories/assets/accessibility.svg
	deleted:    src/stories/assets/addon-library.png
	deleted:    src/stories/assets/assets.png
	deleted:    src/stories/assets/context.png
	deleted:    src/stories/assets/discord.svg
	deleted:    src/stories/assets/docs.png
	deleted:    src/stories/assets/figma-plugin.png
	deleted:    src/stories/assets/github.svg
	deleted:    src/stories/assets/share.png
	deleted:    src/stories/assets/styling.png
	deleted:    src/stories/assets/testing.png
	deleted:    src/stories/assets/theming.png
	deleted:    src/stories/assets/tutorials.svg
	deleted:    src/stories/assets/youtube.svg
	deleted:    src/stories/button.css
	deleted:    src/stories/header.css
	deleted:    src/stories/page.css
